+++ collect-webapp/frontend/src/common/components/DeleteIconButton.js
@@ -5,7 +5,7 @@ import Dialogs from 'common/components/Dialogs'
 import L from 'utils/Labels'
 
 const DeleteIconButton = (props) => {
-  const { onClick: onClickParam, confirmTitle, confirmMessage } = props
+  const { onClick: onClickParam, confirmTitle, confirmMessage, title } = props
 
   const onClick = () => {
     const confirmTitleFinal = confirmTitle || L.l('common.delete.confirm.title')
@@ -17,7 +17,7 @@ const DeleteIconButton = (props) => {
   }
 
   return (
-    <Button onClick={onClick} color="danger" className="btn-delete">
+    <Button onClick={onClick} color="danger" className="btn-delete" title={title}>
       <span className="fa fa-trash" />
     </Button>
   )