+++ collect-webapp/frontend/src/common/components/Dropzone.js
@@ -1,21 +1,28 @@
 import './dropzone.css'
 
 import React from 'react'
+import PropTypes from 'prop-types'
 import { useDropzone } from 'react-dropzone'
 import classNames from 'classnames'
 
 import L from 'utils/Labels'
 
+const Size = {
+  SMALL: 'small',
+  MEDIUM: 'medium',
+  LARGE: 'large',
+}
+
 const Dropzone = (props) => {
   const {
     className: classNameProps,
-    compact,
     handleFileDrop,
     acceptedFileTypes,
     acceptedFileTypesDescription,
     selectedFilePreview,
-    width = '100%',
-    height = '200px',
+    size,
+    width,
+    height,
   } = props
 
   const handleFilesDrop = (acceptedFiles) => {
@@ -38,28 +45,46 @@ const Dropzone = (props) => {
       <input {...getInputProps()} />
       {selectedFilePreview ? (
         <p className="fileDroppedMessage">
-          {!compact && <span className="checked large" />}
+          {size === Size.LARGE && <span className="checked large" />}
           {selectedFilePreview}
         </p>
       ) : (
         <div className="fileSelectContainer">
-          <p>{L.l('forms.dropzone.selectFile')}</p>
-          {!compact && (
+          {size === Size.SMALL ? (
+            <span>{L.l('forms.dropzone.selectFileCompact')}</span>
+          ) : (
+            <p>{L.l('forms.dropzone.selectFile')}</p>
+          )}
+          {size === Size.LARGE && (
             <>
               <br />
               <br />
             </>
           )}
-          <p>{L.l('forms.dropzone.supportedFileTypes', [acceptedFileTypesDescription])}</p>
+          {size !== Size.SMALL && <p>{L.l('forms.dropzone.supportedFileTypes', [acceptedFileTypesDescription])}</p>}
         </div>
       )}
     </div>
   )
 }
 
+Dropzone.propTypes = {
+  className: PropTypes.string,
+  handleFileDrop: PropTypes.func.isRequired,
+  acceptedFileTypes: PropTypes.string.isRequired,
+  acceptedFileTypesDescription: PropTypes.string.isRequired,
+  selectedFilePreview: PropTypes.string,
+  size: PropTypes.oneOf([Size.SMALL, Size.MEDIUM, Size.LARGE]),
+  width: PropTypes.string,
+  height: PropTypes.string,
+}
+
 Dropzone.defaultProps = {
   className: null,
-  compact: false,
+  selectedFilePreview: null,
+  size: 'large',
+  width: '100%',
+  height: '200px',
 }
 
 export default Dropzone